fix: address terminal ai review feedback
- harden terminal AI settings initialization and runtime sync
- fix vendored flutter_gemma review issues and Android compilation
- keep PR branch buildable for review testing

+++ packages/flutter_gemma/lib/core/infrastructure/hnsw_vector_index.dart
@@ -111,7 +111,7 @@ class HnswVectorIndex {
     }
 
     // Over-fetch candidates (HNSW returns approximate results)
-    final candidateCount = (topK * 2).clamp(1, _embeddings.length);
+    final candidateCount = (topK * 2).clamp(1, _embeddings.length).toInt();
     final searchResult = _index!.search(queryEmbedding, candidateCount);
 
     // Recalculate exact similarity and filter by threshold
